Fix some issues with tri and bigram perplexity
Trigram model used to ignore the first token since it is not the last
word of any trigram. Now the model costructs a bigram at the beginning
of the sentence to fix this. Also, bigram used to ignore the first
bigram since its denominator was count(<START>) which was not in the
dictionary. THis has been fixed by using count(<STOP>) as they are the
same.